\name{rite-package}
\alias{rite-package}
\docType{package}
\title{The Right Editor to Write R}
\description{
A simple, but powerful script editor for R, built with tcl/tk. \code{\link{rite}} helps ease new R users' transition to R, a major rite of passage.

\code{\link{rite}} is designed to substitute for the built-in script editor provided by R and improve upon it by offering functionality more commonly found in standalone editors and IDEs (e.g., syntax highlighting, command completion, shortcut keys, find and go-to-line commands, direct acces R help files, etc.), as well as an optional output viewer that eliminates the need to toggle between the script editor and the R console (and that additionally facilitates integration with \code{knitr}.
}
\details{
\tabular{ll}{
Package: \tab rite\cr
Type: \tab Package\cr
Version: \tab 0.2.46\cr
Date: \tab 2013-11-03\cr
License: \tab GPL-2\cr
}
}
\author{
Thomas J. Leeper

Maintainer: Thomas J. Leeper <thosjleeper@gmail.com>
}
%\references{}
\keyword{ package }
%\seealso{}
%\examples{}

