\name{ritesink}
\alias{ritesink}
\title{ritesink}
\description{An experimental output tool}
\usage{
ritesink(new=TRUE)
}
\arguments{
  \item{new}{Ignore.}
}
\details{
The beginnings of an experimental tool for formatting R output. Not really ready for primetime.
}
%\value{}
%\references{}
\author{Thomas J. Leeper}
%\note{}
%\seealso{}
%\examples{\dontrun{ritesink()}}
\keyword{ IO }
